feat: add not_contains operator
