+++ core/src/main/kotlin/com/rapatao/projects/ruleset/engine/types/builder/extensions/ContainsExtensions.kt
@@ -20,3 +20,21 @@ infix fun String.expContains(right: String): Expression = ExpressionBuilder.left
  * @return An [Expression] object representing the containment check, with the operator set to "contains".
  */
 infix fun List<Any>.expContains(right: Any): Expression = ExpressionBuilder.left(this).contains(right)
+
+/**
+ * Creates an expression to check if the current string not contains the specified substring.
+ *
+ * @receiver The string in which to check for the presence of the substring.
+ * @param right The substring to search for within the current string.
+ * @return An [Expression] object representing the containment check, with the operator set to "not_contains".
+ */
+infix fun String.expNotContains(right: String): Expression = ExpressionBuilder.left(this).notContains(right)
+
+/**
+ * Creates an expression to check if the current list not contains the specified element.
+ *
+ * @receiver The list in which to check for the presence of the element.
+ * @param right The element to search for within the current list.
+ * @return An [Expression] object representing the containment check, with the operator set to "not_contains".
+ */
+infix fun List<Any>.expNotContains(right: Any): Expression = ExpressionBuilder.left(this).notContains(right)
