+++ graaljs-evaluator/src/main/kotlin/com/rapatao/projects/ruleset/engine/evaluator/graaljs/operator/Contains.kt
@@ -1,0 +1,19 @@
+package com.rapatao.projects.ruleset.engine.evaluator.graaljs.operator
+
+import com.rapatao.projects.ruleset.engine.context.EvalContext
+import com.rapatao.projects.ruleset.engine.types.operators.ContainsOperator
+
+internal class Contains : ContainsOperator() {
+    override fun process(context: EvalContext, left: Any?, right: Any?): Boolean =
+        context.evaluate(
+            """
+                (function() {
+                    if (Array.isArray(${left})) {
+                        return ${left}.includes(${right})
+                    } else {
+                        return ${left}.indexOf(${right}) !== -1
+                    }
+                })()
+            """.trimIndent()
+        )
+}
