+++ graaljs-evaluator/src/main/kotlin/com/rapatao/projects/ruleset/engine/evaluator/graaljs/operator/GreaterThan.kt
@@ -1,0 +1,9 @@
+package com.rapatao.projects.ruleset.engine.evaluator.graaljs.operator
+
+import com.rapatao.projects.ruleset.engine.context.EvalContext
+import com.rapatao.projects.ruleset.engine.types.operators.GreaterThanOperator
+
+internal class GreaterThan : GreaterThanOperator() {
+    override fun process(context: EvalContext, left: Any?, right: Any?): Boolean =
+        context.evaluate("($left) > ($right)")
+}
