+++ rhino-evaluator/src/main/kotlin/com/rapatao/projects/ruleset/engine/evaluator/rhino/operator/NotContains.kt
@@ -1,0 +1,11 @@
+package com.rapatao.projects.ruleset.engine.evaluator.rhino.operator
+
+import com.rapatao.projects.ruleset.engine.context.EvalContext
+import com.rapatao.projects.ruleset.engine.types.operators.NotContainsOperator
+
+class NotContains : NotContainsOperator() {
+    private val containsOperator = Contains()
+
+    override fun process(context: EvalContext, left: Any?, right: Any?): Boolean =
+        !containsOperator.process(context, left, right)
+}
