+++ rhino-evaluator/src/main/kotlin/com/rapatao/projects/ruleset/engine/evaluator/rhino/operator/NotEquals.kt
@@ -1,0 +1,9 @@
+package com.rapatao.projects.ruleset.engine.evaluator.rhino.operator
+
+import com.rapatao.projects.ruleset.engine.context.EvalContext
+import com.rapatao.projects.ruleset.engine.types.operators.NotEqualsOperator
+
+internal class NotEquals : NotEqualsOperator() {
+    override fun process(context: EvalContext, left: Any?, right: Any?): Boolean =
+        context.evaluate("($left) != ($right)")
+}
